- Add execution test - Refactor diagram components into separate files
