+++ ui-next/src/pages/execution/AgentExecution/agentExecutionDiagram/constants.ts
@@ -1,0 +1,4 @@
+export const EDGE_DEFAULT = "#757575";
+export const EDGE_COMPLETED = "#40BA56";
+export const MIN_ZOOM = 0.1;
+export const MAX_ZOOM = 2.5;
